backend/requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.96.0
anyio==4.13.0
audioread==3.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
comtypes==1.4.16
decorator==5.2.1
distro==1.9.0
docstring_parser==0.18.0
fastapi==0.136.0
filelock==3.28.0
fsspec==2026.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
HyperPyYAML==1.2.3
idna==3.11
Jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
lazy-loader==0.5
librosa==0.11.0
llvmlite==0.47.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
networkx==3.6.1
numba==0.65.0
numpy==2.4.4
openai==2.32.0
packaging==26.1
platformdirs==4.9.6
pooch==1.9.0
pycparser==3.0
pydantic==2.13.2
pydantic_core==2.46.2
Pygments==2.20.0
pypiwin32==223
python-dotenv==1.2.2
python-multipart==0.0.26
pyttsx3==2.99
pywin32==311
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
rich==15.0.0
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentencepiece==0.2.1
setuptools==81.0.0
shellingham==1.5.4
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
speechbrain==1.1.0
starlette==1.0.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
torchaudio==2.11.0
tqdm==4.67.3
transformers==5.5.4
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
websockets==16.0
